requirements.txt
# 向量数据库
# chromadb==0.4.22

# 向量化模型
sentence-transformers==2.2.2
transformers==4.36.2
torch>=2.1.0

# Web框架 (升级到支持pydantic 2.x)
fastapi>=0.104.0
uvicorn>=0.24.0
python-multipart==0.0.6

# 数据验证 (升级到pydantic 2.x)
pydantic>=2.0.0

# 数据库驱动
pymysql==1.1.0
pymongo==4.6.1
elasticsearch==8.11.0

# 数据处理
pandas>=2.2.0
numpy>=1.26.0

# 工具库
python-dotenv==1.0.0
tiktoken==0.5.2
aiohttp>=3.9.0

# LangChain (升级到最新版本)
langchain>=0.1.0
langchain-community>=0.0.10
langchain-openai>=0.0.5

# FAISS
faiss-cpu>=1.7.4 